Face-recognition functions factered out in a new module.
